refactor: remove unnecessary code and dependencies, update scripts
