+++ .gitignore
@@ -13,6 +13,7 @@ coverage
 dist
 dist-ssr
 *.local
+.env
 
 # Editor directories and files
 .vscode/*
@@ -24,3 +25,7 @@ dist-ssr
 *.njsproj
 *.sln
 *.sw?
+
+# CDK asset staging directory
+.cdk.staging
+cdk.out
