--- cdk/.gitignore
@@ -1,9 +1,0 @@
-*.js
-!jest.config.js
-*.d.ts
-node_modules
-.env
-
-# CDK asset staging directory
-.cdk.staging
-cdk.out
