+++ cdk/package.json
@@ -2,30 +2,26 @@
   "name": "cdk",
   "version": "0.1.0",
   "bin": {
-    "cdk": "bin/cdk.js"
+    "cdk": "cdk.js"
   },
   "scripts": {
     "build": "tsc",
     "watch": "tsc -w",
-    "test": "jest",
     "cdk": "cdk",
     "bootstrap": "npm run cdk bootstrap",
     "ls": "npm run cdk ls",
+    "diff": "npm run cdk diff",
     "deploy": "npm run bootstrap && npm run cdk deploy",
-    "destroy": "npm run bootstrap && npm run cdk destroy"
+    "destroy": "npm run cdk destroy"
   },
   "devDependencies": {
-    "@types/jest": "^29.5.12",
     "@types/node": "20.12.7",
-    "jest": "^29.7.0",
-    "ts-jest": "^29.1.2",
     "aws-cdk": "2.144.0",
     "ts-node": "^10.9.2",
     "typescript": "~5.4.5"
   },
   "dependencies": {
     "aws-cdk-lib": "2.144.0",
-    "constructs": "^10.0.0",
-    "source-map-support": "^0.5.21"
+    "constructs": "^10.0.0"
   }
 }
