+++ package.json
@@ -11,9 +11,9 @@
     "test:coverage": "vitest run --coverage",
     "lint": "eslint --ignore-path .gitignore --fix src",
     "prettier": "prettier src --write",
-    "deploy": "cd cdk && npm run deploy",
-    "destroy": "cd cdk && npm run destroy",
-    "build:deploy": "npm run build && npm run deploy"
+    "cdk:deploy": "cd cdk && npm run deploy",
+    "cdk:destroy": "cd cdk && npm run destroy",
+    "build:deploy": "npm run build && npm run cdk:deploy"
   },
   "dependencies": {
     "@emotion/react": "^11.9.3",
